pyproject.toml
[project]
name = "docverify"
version = "1.0.0"
description = "Verified answers from your documentation — with proof."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "pdfplumber>=0.11.0",
    "tiktoken>=0.7.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "transformers>=4.40.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ragas = ["ragas>=0.2.0"]
aegra = ["aegra-cli>=0.1.0"]
dev = ["pytest", "pytest-asyncio", "httpx"]

[project.scripts]
docverify = "docverify.graph:run"
docverify-api = "docverify.server:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
